move fastspec and fastphot scripts to single, new fastspecfit.py module
