+++ bin/fastspec
@@ -5,11 +5,5 @@ Example call:
   fastspec /global/cfs/cdirs/desi/spectro/redux/daily/tiles/80608/20201215/zbest-9-80608-20201215.fits -o fastspec.fits --ntargets 2 --mp 1
 
 """
-import os
-from desiutil.log import get_logger
-from fastspecfit.scripts import fastspec
-
-log = get_logger()
-
-fastspec.main(comm=None)
-
+from fastspecfit.fastspecfit import fastspec
+fastspec(comm=None)
